Moved GalleryFileWidget to a separate file
